feat(forum): refactor NewPostPage to use ClientNewPost component for improved structure and authentication handling
