+++ src/components/features/forum/ForumFeed.component.tsx
@@ -2,7 +2,6 @@
 
 import React, { useState, useEffect, useCallback, useRef } from "react";
 import { useRouter } from "next/navigation";
-import Link from "next/link";
 
 import { supabase } from "@/helper/supabase.helper";
 import { Card } from "@/ui/Card.ui";
